# apps/mealplanr-api/src/swagger.json
{
	"components": {
...
		"schemas": {
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"CategoryDocument": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"Pick_CategoryDocument.name-or-type_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CategoryCreationParams": {
				"$ref": "#/components/schemas/Pick_CategoryDocument.name-or-type_"
			},
			"StoreDocument": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"Pick_StoreDocument.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"StoreCreationParams": {
				"$ref": "#/components/schemas/Pick_StoreDocument.name_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "mealplanr-api",
		"version": "0.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/categories/{categoryId}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDocument"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "resource updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDocument"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryCreationParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "resource deleted successfully"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/categories": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "resource created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDocument"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryCreationParams"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}": {
			"get": {
				"operationId": "GetStore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoreDocument"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateStore",
				"responses": {
					"200": {
						"description": "resource updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoreDocument"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreCreationParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStore",
				"responses": {
					"204": {
						"description": "resource deleted successfully"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/stores": {
			"post": {
				"operationId": "CreateStore",
				"responses": {
					"201": {
						"description": "resource created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoreDocument"
								}
							}
						}
					}
				},
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}